% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ExcelTable}
\alias{ExcelTable}
\title{Export a Formatted Excel Table}
\usage{
ExcelTable(
  table,
  title = "Table.N",
  caption = "haha",
  outfile = outfile,
  openfile = TRUE
)
}
\arguments{
\item{table}{A table or data frame to write to Excel.}

\item{title}{A character title for the Excel column header.}

\item{caption}{A caption to add as a footer.}

\item{outfile}{The full path for the output Excel file.}

\item{openfile}{Logical indicating whether to open the file after saving. Default is TRUE.}
}
\value{
The workbook object (invisible).
}
\description{
Exports a matrix or data frame to an Excel file with formatted headers and footnotes.
}
