% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{To_6Cs_ID}
\alias{To_6Cs_ID}
\title{Convert Participant ID to Six Characters}
\usage{
To_6Cs_ID(idname, dt)
}
\arguments{
\item{idname}{A character string of the ID variable name in the data frame.}

\item{dt}{A data frame containing the ID variable.}
}
\value{
A data frame with modified ID values.
}
\description{
Ensures all participant IDs are six characters long by padding with zeros.
}
